requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
openai
python-multipart
python-dotenv
nltk
requests
httpx
aiofiles
pytest
pytest-asyncio
google-generativeai
# Optional: For enhanced functionality
# chromadb==0.4.18
# faiss-cpu==1.7.4
# sentence-transformers==2.2.2
